+++ level_zero/core/source/cmdqueue/cmdqueue.h
@@ -8,8 +8,7 @@
 #pragma once
 
 #include "level_zero/core/source/device/device.h"
-#include <level_zero/ze_common.h>
-#include <level_zero/ze_fence.h>
+#include <level_zero/ze_api.h>
 
 #include <atomic>
 
@@ -39,7 +38,7 @@ struct CommandQueue : _ze_command_queue_handle_t {
     virtual ze_result_t executeCommands(uint32_t numCommands,
                                         void *phCommands,
                                         ze_fence_handle_t hFence) = 0;
-    virtual ze_result_t synchronize(uint32_t timeout) = 0;
+    virtual ze_result_t synchronize(uint64_t timeout) = 0;
 
     static CommandQueue *create(uint32_t productFamily, Device *device, NEO::CommandStreamReceiver *csr,
                                 const ze_command_queue_desc_t *desc, bool isCopyOnly);
